move eash aspect to different modules for readability
